thai code api get chuyen di va lay thong tin gia ve theo diem trung chuyen

+++ APIDatVe/APIDatVe/Database/KhachHang.cs
@@ -27,7 +27,7 @@ namespace APIDatVe.Database
         public string diachi { get; set; }
         public string sodienthoai { get; set; }
         public string email { get; set; }
-        public Nullable<double> diemtichluy { get; set; }
+        public Nullable<int> diemtichluy { get; set; }
         public string matkhau { get; set; }
         public string gioitinh { get; set; }
